The game volume can now be controlled from the options menu!
Added an audio manager to manage the sfx and music volume levels. the audio manager also sets the volume on all the sounds in the game.

+++ SpaceShooter/Asteroid.cpp
@@ -30,7 +30,7 @@ void Asteroid::spawn(sf::Vector2f targetPosition, sf::FloatRect gameBounds)
 	moveDirection /= magnitude;
 }
 
-void Asteroid::increaseSpeed(float amount)
+void Asteroid::increaseSpeed(int amount)
 {
 	moveSpeed += amount;
 }
